Fix dependency confusion issues
If all rubygems sources implement dependency API endpoints, and there's
no multiple global rubygems sources, we can avoid the need for an
aggregate rubygems source and instead compute all source requirements
(including for indirect dependencies) explicitly before resolution.

Note that with this fix, once an indirect dependency is present in a
scoped source, it will be picked up from there instead of from the
global source, even if the version present in the global source
satisfies requirements whereas the version in the scoped source does
not. It doesn't seem like a big deal, and the way I see it, it's the
less surprising behavior.